requirements.txt
anthropic==0.28.0
openai>=1.0.0
beautifulsoup4==4.12.2
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
lxml==4.9.3
markdown==3.5.1
python-pptx>=0.6.21
Pillow>=9.0.0
pyyaml>=6.0
google-genai>=1.0.0
httpx>=0.24.0